moved base model class to a common package
