+++ crm-api/src/main/kotlin/com/james/crm/api/core/exception/GlobalExceptionHandler.kt
@@ -8,7 +8,10 @@
 package com.james.crm.api.core.exception
 
 import com.fasterxml.jackson.databind.JsonMappingException
-import org.springframework.http.HttpStatus
+import com.james.crm.api.core.common.ApiResponse
+import com.james.crm.api.core.common.ErrorApiResponse
+import org.springframework.http.HttpStatus.BAD_REQUEST
+import org.springframework.http.HttpStatus.METHOD_NOT_ALLOWED
 import org.springframework.http.ResponseEntity
 import org.springframework.http.converter.HttpMessageNotReadableException
 import org.springframework.validation.FieldError
@@ -24,8 +27,8 @@ import java.util.*
 class GlobalExceptionHandler {
 
     @ExceptionHandler(MethodArgumentNotValidException::class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    fun handleValidationExceptions(ex: MethodArgumentNotValidException): ResponseEntity<Map<String, String>> {
+    @ResponseStatus(BAD_REQUEST)
+    fun handleValidationExceptions(ex: MethodArgumentNotValidException): ResponseEntity<ApiResponse<Nothing>> {
         val resolve = mutableMapOf<String, String>()
         ex.bindingResult.allErrors.forEach { error ->
             if (error is FieldError) {
@@ -34,32 +37,46 @@ class GlobalExceptionHandler {
                 resolve[error.objectName] = error.defaultMessage ?: "failed validation"
             }
         }
-        return ResponseEntity.badRequest().body(resolve)
+        val response = ErrorApiResponse(
+            message = BAD_REQUEST.reasonPhrase,
+            status = BAD_REQUEST.value(),
+            errors = resolve.map { entry: Map.Entry<String, String> -> "${entry.key} -> ${entry.value}" }
+        )
+        return ResponseEntity.badRequest().body(response)
     }
 
     @ExceptionHandler(HttpMessageNotReadableException::class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    fun handleHttpMessageNotReadableException(ex: HttpMessageNotReadableException): ResponseEntity<String> {
+    @ResponseStatus(BAD_REQUEST)
+    fun handleHttpMessageNotReadableException(ex: HttpMessageNotReadableException): ResponseEntity<ErrorApiResponse> {
         if (ex.cause is JsonMappingException) {
             val jsonMappingException = ex.cause as JsonMappingException
             val fieldName = jsonMappingException.path.joinToString(".") { it.fieldName }
             val errorMessage = "Invalid value or format for field: '$fieldName'."
-            return ResponseEntity.badRequest().body(errorMessage)
+
+            val response = ErrorApiResponse(
+                message = BAD_REQUEST.reasonPhrase,
+                status = BAD_REQUEST.value(), errors = listOf(errorMessage)
+            )
+            return ResponseEntity.status(BAD_REQUEST).body(response)
         }
 
         // Default error message if the cause is not a JsonMappingException
         val defaultErrorMessage = "Malformed JSON request or invalid request body format."
-        return ResponseEntity.badRequest().body(defaultErrorMessage)
+        val response = ErrorApiResponse(
+            message = BAD_REQUEST.reasonPhrase,
+            status = BAD_REQUEST.value(),
+            errors = listOf(defaultErrorMessage)
+        )
+        return ResponseEntity.badRequest().body(response)
     }
 
-
     @ExceptionHandler(HttpRequestMethodNotSupportedException::class)
-    fun handleError(e: HttpRequestMethodNotSupportedException): ResponseEntity<Map<String, String>>? {
-        val response: MutableMap<String, String> = HashMap()
-        response["status"] = false.toString()
-        response["code"] = HttpStatus.METHOD_NOT_ALLOWED.reasonPhrase
-        response["message"] = "It seems you're using the wrong HTTP method"
-        response["errors"] = "${e.message}"
-        return ResponseEntity.status(HttpStatus.METHOD_NOT_ALLOWED).body(response)
+    fun handleError(e: HttpRequestMethodNotSupportedException): ResponseEntity<ErrorApiResponse> {
+        val apiErrorResponse = ErrorApiResponse(
+            message = METHOD_NOT_ALLOWED.reasonPhrase,
+            status = METHOD_NOT_ALLOWED.value(),
+            errors = listOf("It seems you're using the wrong HTTP method: ${e.message}")
+        )
+        return ResponseEntity.status(METHOD_NOT_ALLOWED).body(apiErrorResponse)
     }
 }
