+++ crm-api/src/main/kotlin/com/james/crm/api/modules/people/data/dto/AgentDto.kt
@@ -8,7 +8,7 @@
 package com.james.crm.api.modules.people.data.dto
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.james.crm.api.core.model.Mapper
+import com.james.crm.api.core.common.Mapper
 import com.james.crm.api.modules.people.domain.enums.UserTypeEnum
 import com.james.crm.api.modules.people.domain.model.Agent
 import com.james.crm.api.modules.team.data.dto.TeamDto
