+++ crm-api/src/main/kotlin/com/james/crm/api/modules/people/data/dto/ManagerDto.kt
@@ -1,7 +1,7 @@
 package com.james.crm.api.modules.people.data.dto
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.james.crm.api.core.model.Mapper
+import com.james.crm.api.core.common.Mapper
 import com.james.crm.api.modules.people.domain.enums.UserTypeEnum
 import com.james.crm.api.modules.people.domain.model.Agent
 import com.james.crm.api.modules.people.domain.model.Manager
