+++ crm-api/src/main/kotlin/com/james/crm/api/modules/people/domain/model/Manager.kt
@@ -7,8 +7,8 @@
 
 package com.james.crm.api.modules.people.domain.model
 
+import com.james.crm.api.core.common.Base
 import com.james.crm.api.core.constant.DatabaseTable.Companion.MANAGER
-import com.james.crm.api.core.model.Base
 import com.james.crm.api.modules.people.domain.model.submodel.Contact
 import com.james.crm.api.modules.people.domain.model.submodel.Profile
 import jakarta.persistence.*
