+++ crm-api/src/main/kotlin/com/james/crm/api/modules/people/domain/model/submodel/Location.kt
@@ -7,8 +7,8 @@
 
 package com.james.crm.api.modules.people.domain.model.submodel
 
+import com.james.crm.api.core.common.Base
 import com.james.crm.api.core.constant.DatabaseTable.Companion.LOCATION
-import com.james.crm.api.core.model.Base
 import jakarta.persistence.Entity
 import jakarta.persistence.Table
