+++ crm-api/src/main/kotlin/com/james/crm/api/modules/people/domain/model/submodel/Resource.kt
@@ -7,9 +7,8 @@
 
 package com.james.crm.api.modules.people.domain.model.submodel
 
-import com.james.crm.api.core.constant.DatabaseTable
+import com.james.crm.api.core.common.Base
 import com.james.crm.api.core.constant.DatabaseTable.Companion.RESOURCE
-import com.james.crm.api.core.model.Base
 import com.james.crm.api.modules.people.domain.model.User
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
@@ -18,7 +17,7 @@ import jakarta.persistence.Table
 
 @Entity
 @Table(name = RESOURCE)
-class Resource: Base() {
+class Resource : Base() {
     var title: String = ""
     var type: String = ""
     var url: String = ""
