+++ crm-api/src/main/kotlin/com/james/crm/api/modules/team/data/usecase/contract/ISetTeamTaskUseCase.kt
@@ -7,7 +7,7 @@
 
 package com.james.crm.api.modules.team.data.usecase.contract
 
-import com.james.crm.api.core.model.Usecase
+import com.james.crm.api.core.common.Usecase
 import com.james.crm.api.modules.team.data.dto.TeamDto
 
 interface ISetTeamTaskUseCase : Usecase<Pair<String, String>, TeamDto>
